requirements.txt
transparent-background
gradio
opencv-python
numpy
Pillow
torch
torchvision
fastapi
uvicorn
python-multipart
